Bases_de_datos_Entrega_1/requirements.txt
python-dotenv

pandas
xlrd

mysql-connector-python

streamlit
watchdog